requirements.txt
pyspark==2.4.4
python-dateutil==2.7.3
requests==2.19.1